+++ lib/notion/getNotionData.js
@@ -3,7 +3,8 @@ import { getDataFromCache, setDataToCache } from '@/lib/cache/cache_manager'
 import { getPostBlocks } from '@/lib/notion/getPostBlocks'
 import { idToUuid } from 'notion-utils'
 import { defaultMapImageUrl } from 'react-notion-x'
-import { deepClone, isIterable } from '../utils'
+import { deepClone } from '../utils'
+import { getAllCategories } from './getAllCategories'
 import getAllPageIds from './getAllPageIds'
 import { getAllTags } from './getAllTags'
 import getPageProperties from './getPageProperties'
@@ -31,8 +32,6 @@ export async function getGlobalNotionData({
   delete notionPageData.schema
   delete notionPageData.rawMetadata
   delete notionPageData.pageIds
-  delete notionPageData.tagOptions
-  delete notionPageData.categoryOptions
   return notionPageData
 }
 
@@ -115,46 +114,6 @@ function getCategoryOptions(schema) {
   if (!schema) return {}
   const categorySchema = Object.values(schema).find(e => e.name === BLOG.NOTION_PROPERTY_NAME.category)
   return categorySchema?.options || []
-}
-
-/**
- * 获取所有文章的分类
- * @param allPosts
- * @returns {Promise<{}|*[]>}
- */
-function getAllCategories({ allPages, categoryOptions, sliceCount = 0 }) {
-  const allPosts = allPages.filter(page => page.type === 'Post')
-  if (!allPosts || !categoryOptions) {
-    return []
-  }
-  // 计数
-  let categories = allPosts.map(p => p.category)
-  categories = [...categories.flat()]
-  const categoryObj = {}
-  categories.forEach(category => {
-    if (category in categoryObj) {
-      categoryObj[category]++
-    } else {
-      categoryObj[category] = 1
-    }
-  })
-  const list = []
-  if (isIterable(categoryOptions)) {
-    for (const c of categoryOptions) {
-      const count = categoryObj[c.value]
-      if (count) {
-        list.push({ id: c.id, name: c.value, color: c.color, count })
-      }
-    }
-  }
-
-  // 按照数量排序
-  // list.sort((a, b) => b.count - a.count)
-  if (sliceCount && sliceCount > 0) {
-    return list.slice(0, sliceCount)
-  } else {
-    return list
-  }
 }
 
 /**
@@ -279,7 +238,7 @@ async function getPageRecordMapByNotionAPI({ pageId, from }) {
 
   const customNav = getCustomNav({ allPages })
   const categories = getAllCategories({ allPages, categoryOptions, sliceCount: BLOG.PREVIEW_CATEGORY_COUNT })
-  const tags = getAllTags({ allPages, tagOptions })
+  const tags = getAllTags({ allPages, sliceCount: BLOG.PREVIEW_TAG_COUNT, tagOptions })
   const latestPosts = getLatestPosts({ allPages, from, latestPostCount: 5 })
 
   return {
