+++ pages/archive/index.js
@@ -20,10 +20,10 @@ const ArchiveIndex = props => {
 
 export async function getStaticProps() {
   const props = await getGlobalNotionData({ from: 'archive-index' })
-  const { allPages } = props
-  const allPosts = allPages.filter(page => page.type === 'Post' && page.status === 'Published')
   // 处理分页
-  props.posts = allPosts
+  props.posts = props.allPages.filter(page => page.type === 'Post' && page.status === 'Published')
+  delete props.allPages
+
   return {
     props,
     revalidate: 1
