+++ pages/category/index.js
@@ -2,6 +2,7 @@ import { getGlobalNotionData } from '@/lib/notion/getNotionData'
 import React from 'react'
 import { useGlobal } from '@/lib/global'
 import * as ThemeMap from '@/themes'
+import { getAllCategories } from '@/lib/notion/getAllCategories'
 
 /**
  * 分类首页
@@ -24,10 +25,9 @@ export default function Category(props) {
 }
 
 export async function getStaticProps() {
-  const props = await getGlobalNotionData({
-    from: 'category-index-props',
-    categoryCount: 0
-  })
+  const props = await getGlobalNotionData({ from: 'category-index-props' })
+  props.categories = getAllCategories({ allPages: props.allPages, categoryOptions: props.categoryOptions, sliceCount: 0 })
+  delete props.categoryOptions
   return {
     props,
     revalidate: 1
