+++ pages/tag/index.js
@@ -2,6 +2,7 @@ import { getGlobalNotionData } from '@/lib/notion/getNotionData'
 import React from 'react'
 import { useGlobal } from '@/lib/global'
 import * as ThemeMap from '@/themes'
+import { getAllTags } from '@/lib/notion'
 
 /**
  * 标签首页
@@ -26,6 +27,9 @@ const TagIndex = props => {
 export async function getStaticProps() {
   const from = 'tag-index-props'
   const props = await getGlobalNotionData({ from })
+  console.log('获取所有标签', props)
+  props.tags = getAllTags({ allPages: props.allPages, sliceCount: 0, tagOptions: props.tagOptions })
+  delete props.tagOptions
   return {
     props,
     revalidate: 1
