+++ themes/next/LayoutCategory.js
@@ -2,15 +2,20 @@ import LayoutBase from './LayoutBase'
 import StickyBar from './components/StickyBar'
 import CategoryList from './components/CategoryList'
 import BlogPostListScroll from './components/BlogPostListScroll'
+import BlogPostListPage from './components/BlogPostListPage'
+import BLOG from '@/blog.config'
 
 export const LayoutCategory = (props) => {
-  const { tags, posts, category, categories } = props
+  const { category, categories } = props
   return <LayoutBase currentCategory={category} {...props}>
     <StickyBar>
       <CategoryList currentCategory={category} categories={categories} />
     </StickyBar>
     <div className='md:mt-8'>
-      <BlogPostListScroll posts={posts} tags={tags} currentCategory={category}/>
+      {BLOG.POST_LIST_STYLE !== 'page'
+        ? <BlogPostListScroll {...props} showSummary={true} />
+        : <BlogPostListPage {...props} />
+    }
     </div>
   </LayoutBase>
 }
