+++ themes/next/LayoutTag.js
@@ -2,16 +2,21 @@ import LayoutBase from './LayoutBase'
 import StickyBar from './components/StickyBar'
 import TagList from './components/TagList'
 import BlogPostListScroll from './components/BlogPostListScroll'
+import BlogPostListPage from './components/BlogPostListPage'
+import BLOG from '@/blog.config'
 
 export const LayoutTag = (props) => {
-  const { tags, posts, tag } = props
+  const { tags, tag } = props
 
   return <LayoutBase currentTag={tag} {...props}>
     <StickyBar>
       <TagList tags={tags} currentTag={tag}/>
     </StickyBar>
     <div className='md:mt-8'>
-      <BlogPostListScroll posts={posts} tags={tags} currentTag={tag}/>
+      {BLOG.POST_LIST_STYLE !== 'page'
+        ? <BlogPostListScroll {...props} showSummary={true} />
+        : <BlogPostListPage {...props} />
+    }
     </div>
   </LayoutBase>
 }
